Python_Services/requirements.txt
fastapi
uvicorn[standard]
pandas
pymongo[srv]
python-dotenv
scikit-learn